- Move bloglist to liveblog.blogs - Clean bloglist code - Update .gitignore - Add app_settings to prevent circular imports
