+++ server/app.py
@@ -10,20 +10,23 @@
 # at https://www.sourcefabric.org/superdesk/license
 
 
+import os
+
 import jinja2
-from flask_cache import Cache
-from liveblog.common import BlogCache
+
 import flask_s3
+import settings
+from flask_cache import Cache
+from liveblog.analytics.analytics import analytics_blueprint
+from liveblog.blogs.blogslist import (bloglist_assets_blueprint,
+                                      bloglist_blueprint)
 from liveblog.blogs.embeds import embed_blueprint
+from liveblog.common import BlogCache
+from liveblog.marketplace.marketer import marketers_blueprint
+from liveblog.syndication.blogs import \
+    blogs_blueprint as syndication_blogs_blueprint
 from liveblog.syndication.producer import producers_blueprint
 from liveblog.syndication.syndication import syndication_blueprint
-from liveblog.syndication.blogs import blogs_blueprint as syndication_blogs_blueprint
-from liveblog.marketplace.marketer import marketers_blueprint
-
-from liveblog.analytics.analytics import analytics_blueprint
-
-import os
-import settings
 from superdesk.factory import get_app as superdesk_app
 
 
@@ -70,6 +73,12 @@ def get_app(config=None):
 
     # Embed feature.
     app.register_blueprint(embed_blueprint)
+
+    # Embed bloglist.
+    app.register_blueprint(bloglist_assets_blueprint)
+    app.register_blueprint(bloglist_blueprint)
+
+    # Analytics.
     app.register_blueprint(analytics_blueprint)
 
     # Syndication feature.
