+++ server/liveblog/blogs/tasks.py
@@ -1,6 +1,8 @@
 import io
 import logging
+import os
 
+import magic
 import superdesk
 from celery.exceptions import SoftTimeLimitExceeded
 from flask import current_app as app
@@ -9,22 +11,24 @@ from superdesk import get_resource_service
 from superdesk.celery_app import celery
 from superdesk.notification import push_notification
 
+from .app_settings import (BLOGLIST_ASSETS, BLOGSLIST_ASSETS_DIR,
+                           BLOGSLIST_DIRECTORY, CONTENT_TYPES)
 from .embeds import embed
 from .exceptions import MediaStorageUnsupportedForBlogPublishing
-from .utils import check_media_storage, get_blog_path
+from .utils import check_media_storage, get_blog_path, get_bloglist_path
 
 logger = logging.getLogger('superdesk')
 
 
 def publish_embed(blog_id, api_host=None, theme=None):
-    # Get html using embed() blueprint
+    # Get html using embed() blueprint.
     html = embed(blog_id, api_host, theme)
     check_media_storage()
     file_path = get_blog_path(blog_id)
-    # Remove existing file
+    # Remove existing file.
     app.media.delete(app.media.media_id(file_path, version=False))
     logger.warning('Embed file "{}" for blog "{}" removed from s3'.format(file_path, blog_id))
-    # Upload new file
+    # Upload new file.
     file_id = app.media.put(io.BytesIO(bytes(html, 'utf-8')), filename=file_path, content_type='text/html',
                             version=False)
     logger.warning('Embed file "{}" for blog "{}" uploaded to s3'.format(file_path, blog_id))
@@ -34,7 +38,7 @@ def publish_embed(blog_id, api_host=None, theme=None):
 def delete_embed(blog_id):
     check_media_storage()
     file_path = get_blog_path(blog_id)
-    # Remove existing file
+    # Remove existing file.
     app.media.delete(file_path)
 
 
@@ -82,3 +86,76 @@ def delete_blog_embed_on_s3(self, blog_id, safe=True):
         raise self.retry(exc=e, max_retries=S3_CELERY_MAX_RETRIES, countdown=S3_CELERY_COUNTDOWN)
     finally:
         logger.warning('delete_blog_on_s3 for blog "{}" finished.'.format(blog_id))
+
+
+@celery.task(soft_time_limit=1800)
+def publish_assets(asset_type):
+    # TODO: add retry
+    assets = BLOGLIST_ASSETS.copy()
+    # version_path = os.path.join(BLOGSLIST_DIRECTORY, BLOGSLIST_ASSETS_DIR, assets['version'])
+    # # loads version json from file
+    # version = json.loads(open(version_path, 'rb').read()).get('version', '0.0.0')
+    # Save the file in the media storage if needed
+    for name in (assets[asset_type]):
+        asset_file = os.path.join(BLOGSLIST_DIRECTORY, BLOGSLIST_ASSETS_DIR, name)
+        with open(asset_file, 'rb') as file:
+            # Set the content type.
+            mime = magic.Magic(mime=True)
+            content_type = mime.from_file(asset_file)
+            if content_type == 'text/plain' and name.endswith(tuple(CONTENT_TYPES.keys())):
+                content_type = CONTENT_TYPES[os.path.splitext(name)[1]]
+
+            final_file_name = os.path.join(BLOGSLIST_ASSETS_DIR, name)
+            # Remove existing first.
+            app.media.delete(app.media.media_id(final_file_name, content_type=content_type))
+            # Upload.
+            app.media.put(file.read(), filename=final_file_name, content_type=content_type)
+
+
+@celery.task(soft_time_limit=1800)
+def publish_bloglist_embed_on_s3():
+    # TODO: add retry
+    from .blogslist import render_bloglist_embed # To prevent circular import. TODO: move to .embeds
+    if type(app.media).__name__ is not 'AmazonMediaStorage':
+        pass
+    else:
+        assets = BLOGLIST_ASSETS.copy()
+
+        # Publish version file to get the asset_root.
+        version_file = os.path.join(BLOGSLIST_ASSETS_DIR, assets.get('version'))
+
+        # Remove existing first.
+        app.media.delete(app.media.media_id(version_file, content_type='application/json'))
+
+        # Upload to Amazon S3.
+        bloglist_path = os.path.join(BLOGSLIST_DIRECTORY, version_file)
+        with open(bloglist_path, 'rb') as f:
+            file_id = app.media.put(f.read(), filename=version_file, content_type='application/json')
+
+        assets_public_url = superdesk.upload.url_for_media(file_id)
+
+        # Correct assets public url path.
+        assets_public_url = assets_public_url.replace(assets['version'], '')
+        assets_public_url = assets_public_url.replace('http://', '//')
+        html = render_bloglist_embed(assets_root=assets_public_url)
+        file_path = get_bloglist_path()
+
+        # Eemove existing.
+        app.media.delete(app.media.media_id(file_path, version=False))
+        # upload
+        file_id = app.media.put(io.BytesIO(bytes(html, 'utf-8')),
+                                filename=file_path,
+                                content_type='text/html',
+                                version=False)
+        public_url = superdesk.upload.url_for_media(file_id)
+
+        # Retrieves all opened blogs.
+        blogslist_service = get_resource_service('blogslist')
+        for blogslist in blogslist_service.get(req=None, lookup=dict(key='bloglist')):
+            get_resource_service('blogslist').system_update(blogslist['_id'], {'value': public_url}, blogslist)
+        else:
+            blogslist_service.create([{'key': 'bloglist', 'value': public_url}])
+
+        # publish_assets('template')
+        publish_assets('scripts')
+        publish_assets('styles')
