+++ server/liveblog/blogs/templates/blog-list-embed.html
