+++ server/liveblog/blogs/utils.py
@@ -12,5 +12,9 @@ def get_blog_path(blog_id):
     return 'blogs/%s/index.html' % (blog_id)
 
 
+def get_bloglist_path():
+    return os.path.join(BLOGSLIST_ASSETS_DIR, 'index.html')
+
+
 def is_relative_to_current_folder(url):
     return not (url.startswith('/') or url.startswith('http://') or url.startswith('https://'))
